pyproject.toml
[project]
name = "amplifier-connector-slack"
version = "0.1.0"
description = "Slack connector for Amplifier — bridge Slack messages to Amplifier sessions via Socket Mode"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "amplifier-foundation",
    "slack-bolt>=1.18",
    "aiohttp>=3.9",
    "click>=8.0",
    "python-dotenv>=1.0",
]

[project.scripts]
slack-connector = "slack_connector.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/slack_connector"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pyright]
pythonVersion = "3.11"
strict = false
